Implement Applier functions
Change the NewApplier and Reset signatures to take a commit instead of a
tree, since we need to the commit SHA to use as a parent when creating a
new commit. This is also closer to how people think about apply patches.

There are still a few TODOs, mostly around error reporting and edge
cases. I'll clean these up in a future commit.