Always create blobs for tree entries
The tree creation API only accepts string content, which doesn't work
for binary files. Instead of leaving "eager blobs" as an option that
only works for text files and detecting the file type, always create
blobs for entries. If it turns out that these API calls are a
significant part of the apply time for large text patches, I can
consider adding an option again.