Add integration tests for patch application (#6)
Because this library is pretty tightly coupled with GitHub, the tests
require a GitHub token and use the real API. If the required values are
not in the environment, tests that use them are skipped.

Tests are loaded dynamically by looking for patch files in the testdata
directory. Each test applies the patch against fixed base content and
then compares the result with the expected files.