Implement GraphQLApplier
GitHub recently added the createCommitOnBranch mutation, which
simplifies commit creation and adds free signing support, with some
limitations that don't make it obviously better than the old approach in
all cases.

Due to these downsides, the different contract, and to support GitHub
Enterprise, implement this as a new type instead of modifying the
existing applier.