Implement GraphQLApplier (#15)
GitHub recently added the createCommitOnBranch mutation, which
simplifies commit creation and adds free signing support, with some
limitations that don't make it obviously better than the old approach in
all cases.

Due to these downsides, the different contract, and to support GitHub
Enterprise, implement this as a new type instead of modifying the
existing applier.

Expand testing to cover the new applier, including patches that it does
not support. Add the new IsUnsupported function to test if apply errors
are due to GraphQL limitations. There's some possibly inefficient code
that gets file modes to check for these limitations, but I don't think
it's worth optimizing at this point.