Add the -fork flag to submit PRs from forks (#77)
Add a new -fork flag to the CLI which will submit a pull request from a
fork, creating it if it does not exist. This avoids needing write access
on the target repository. Also add the -fork-repository flag, which sets
a custom name and owner for the fork. The default is to create a fork
with the same name as the target, owned by the current user.

Last, modify the error handler to print a hint on any 404 error that the
problem might be due to permissions. It should maybe print the same
message on 403 errors, but I haven't seen an example of that yet.