Support reading multiple patches from an mbox file
Using a combined mbox file may be more convenient than providing
multiple files or running the command multiple times, depending on the
source of the patches.

Note that this doesn't do any unescaping of escaped "From " lines in the
message content. This isn't a problem for the actual patch content, but
could lead to unexpected commit messages.